% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{col.grad}
\alias{col.grad}
\title{Gradient Color Palette}
\usage{
col.grad()
}
\value{
Vector of colors to replace default gradient color scale.
}
\description{
Returns 12 colors from the viridis color palette.
}
\examples{

col.grad()

}
