% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{col.univ}
\alias{col.univ}
\title{Universal Color Palette}
\usage{
col.univ()
}
\value{
Vector of colors to replace default discrete color scale.
}
\description{
Combines the npg, aaas, and lancet ggsci palettes for use with datasets
containing up to 29 groups.
}
\examples{

col.univ()

}
