% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2.params.R
\name{Create.proc.params}
\alias{Create.proc.params}
\title{Universal Color Palette}
\usage{
Create.proc.params(d.path, study.md)
}
\arguments{
\item{d.path}{Character string indicating the path to a set of CellRanger files for data processing.}

\item{study.md}{A data frame containing sample metadata variables specific to an individual study.}
}
\value{
Data frame containing a list of parameters to use for scRNA-Seq data processing by Seurat.
}
\description{
Combines the npg, aaas, and lancet ggsci palettes for use with datasets
containing up to 29 groups.
}
\examples{

# Dataset input parameters
list.params <- Create.proc.param(
  "data/",
  data.frame(
    # ID column name (splits by underscore, should be listed first in folder name as in 's01_1_KO_a')
    "Code" = unlist(lapply(strsplit(basename(list.files("data/")),"_",fixed = T),"[",1)),
    # 1st metadata column (include in CellRanger folder name)
    "Knockout" = as.factor(ifelse(grepl("NG",basename(list.files("data/"))),"ctrl","KO")),
    # 2nd metadata column
    "Region" = as.factor(ifelse(grepl("LAE",basename(list.files("data/"))),"1","2")),
    # 3rd metadata column (add/remove columns as needed)
    "Time" = as.factor(ifelse(grepl("D28",basename(list.files("data/"))),"a","b"))
  )
)

}
