% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.6.recluster.R
\name{fun.marker.gene.recluster}
\alias{fun.marker.gene.recluster}
\title{Reclustering Analysis}
\usage{
fun.marker.gene.recluster(so, ct, g.list, md.list)
}
\arguments{
\item{so}{An object of class Seurat. Must contain the columns 'CellType' and 'CellGroup' in the metadata slot.}

\item{ct}{Character string vector containing the name(s) of up to 2 cell type groups present in the CellGroup column.
The provided Seurat object is then subsetted to include only clusters from the provided group names.}

\item{g.list}{Path to an existing text file containing genes for plotting expression.}

\item{md.list}{A vector of character strings indicating metadata columns for overlaying on a loadings plot.}

\item{h.h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs.c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs.r}{Numeric value for row fontsize (passed to ComplexHeatmap).}
}
\value{
A ComplexHeatmap object containing a top-10 marker gene heatmap.
}
\description{
Performs reclustering and basic visualization of a subsetted Seurat Object. Subsets are determined based on CellGroup.
}
\examples{

p.umap <- sc.top10.marker.heatmap(d.annotated,"seurat.clusters",18,24,6,8)

}
