% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.6.format.seurat.R
\name{pc_create_ad}
\alias{pc_create_ad}
\title{Create AnnData from Seurat}
\usage{
pc_create_ad(type1 = "norm", so_list, asy = "PC", mcc = 8)
}
\arguments{
\item{type1}{Data type; provide a character string indicating
if the data have been normalized.}

\item{so_list}{Input Seurat object list.}

\item{asy}{Assay to use for file conversion.}

\item{mcc}{Cores to use for file conversion.}
}
\value{
A h5ad file with the same name as the input object.
}
\description{
Converts a Seurat Object to a h5ad file.
}
\examples{

# ld_seurat <- pc_create_ad(
#   so = readRDS("data/seuratobject.rds")
# )

}
