% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.6.format.seurat.R
\name{pc_create_seurat}
\alias{pc_create_seurat}
\title{Create Seurat Object from QC Dataframe}
\usage{
pc_create_seurat(df, samp_id = "Code", mcc = 2)
}
\arguments{
\item{df}{Input data frame.}

\item{samp_id}{Column containing sample IDs for selected data.}

\item{mcc}{Cores to use for processing input data.}
}
\value{
List of Seurat objects containing filtered phenocycler data.
}
\description{
Converts a filtered phenocycler data frame from pc_qc()
to a Seurat object.
}
\examples{

# ld_seurat <- pc_create_seurat(
#   df = pheno_qc[["data_ind"]]
# )

}
