% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5.phenocycler.QC.R
\name{pc_norm}
\alias{pc_norm}
\title{Phenocycler Normalization}
\usage{
pc_norm(so, mdn, mtd = "slo", list_ch, span1 = 0.05, mcc = 2, its = 3)
}
\arguments{
\item{so}{Input Seurat object.}

\item{mdn}{Number of metadata columns in the input Seurat object.}

\item{mtd}{Normalization method to use (either "zsc" or "slo").}

\item{list_ch}{Vector names of channels to normalize.}

\item{span1}{LOESS span to use (expressed as a proportion of the total cells
present in the dataset). Use a proportion of 0.1 or lower for highly
heterogenous tissues such as lung.}

\item{mcc}{Number of cores to use for sLOESS calculation.}

\item{its}{Number of iterations for fine tuning sLOESS model.}
}
\value{
A normalized Seurat Object for subsequent clustering
and dimension reduction.
}
\description{
Z-score or sparse implementation of LOESS normalization for correcting
gradient artifacts in phenocycler data. Uses method described
by Stevens et. al. 2025 (doi: 10.1038/s41467-025-58135-4).
}
\examples{

# d <- pc_norm(
#   so = ld_ser[[1]],
#   mdn = 9,
#   list_ch = ch_filt
# )

}
