% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5.phenocycler.QC.R
\name{pc_qc}
\alias{pc_qc}
\title{Phenocycler QC}
\usage{
pc_qc(
  qc_type = "all",
  ld,
  col_nums1,
  md_var,
  md_num = 9,
  samp_id = "Code",
  g_col = NULL,
  g_col2 = NULL,
  ch_list,
  col_nums2 = NULL,
  ch_nuc = "DAPI",
  loess_norm = FALSE
)
}
\arguments{
\item{qc_type}{Specific QC steps to perform (
select from "all" or "sum_only")}

\item{ld}{A list of phenocycler segmentation results containing summary data
and segmented cells.}

\item{col_nums1}{Vector of column numbers to include from segmentation
summary data.}

\item{md_var}{Vector for renaming summary data column names.}

\item{md_num}{Number of metadata columns in exported cell segmentation
results. Only used if qc_type is "all".}

\item{samp_id}{Sample ID column name.}

\item{g_col}{(optional) Add a column providing grouping information.}

\item{g_col2}{(optional) Add a column providing grouping information.}

\item{ch_list}{List of channels exported with segmentation results. Note
that the channel list must match the order of channels included within
the results.}

\item{col_nums2}{If qc_type is "all", indicate the column number to
use as the sample ID as well as the X and Y coordinates
for formatting the segmentation results table.}

\item{ch_nuc}{Nuclei channel name.}

\item{loess_norm}{Normalize signal intensities by performing LOESS
prior to z-score scaling.}
}
\value{
List containing QC summary statistics and plots.
}
\description{
Runs quality control of segmented phenocycler data for all samples
}
\examples{

# pheno_data <- pc_qc(
#   ld = ld1,
#   col_nums1 = c(4, 10, 12, 13),
#   md_var = c("Area.um2", "Cells.raw", "Description"),
#   samp_id = "Name",
#   g_col = "Name",
#   g_col2 = "Parent",
#   ch_list = s3,
#   col_nums2 = c(6, 8, 9)
# )

}
