% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5.phenocycler.QC.R
\name{pc_qc_box}
\alias{pc_qc_box}
\title{Phenocycler QC Boxplot}
\usage{
pc_qc_box(df, ch_list = NULL, ch_col = "Channel", int_col = NULL)
}
\arguments{
\item{df}{Input data frame containing signal intensities.}

\item{ch_list}{Channel list for specifying plot order.}

\item{ch_col}{Channel column for spliting individual boxplots. Used to
specify the x-axis and can alternatively plot a different variable if
individual channel intensities are not available}

\item{int_col}{Intensity column to use.}
}
\value{
Boxplot of individual channel or summary of signal intensities.
}
\description{
Generates boxplot of signal intensity
}
\examples{

# pheno_qc_box <- pc_qc_box(
#   df = ld1,
#   ch_list = s3[[1]]
#   int_col = "value.z"
# )

}
