% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5.phenocycler.QC.R
\name{pc_qc_ch}
\alias{pc_qc_ch}
\title{Phenocycler Channel QC}
\usage{
pc_qc_ch(
  type1 = "img",
  so,
  samp_id = "Code",
  ch_id,
  col1 = col_grad(scm = 4),
  loess_smooth = FALSE
)
}
\arguments{
\item{so}{Input Seurat object.}

\item{samp_id}{Sample ID column.}

\item{ch_id}{Channel name.}

\item{col1}{Color scheme to use.}

\item{type}{Either "img" or "sum."}
}
\value{
Panel containing cell map and axis intensity scatter plots
for a specified channel.
}
\description{
Generates scatterplots of segmented cell signal intensity; Must have
X and Y coordinates available in the input dataframe.
}
\examples{

# d <- pc_qc_ch(
#   ld_ser[[1]],
#   ch_id = "DAPI"
# )

}
