% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5.phenocycler.QC.R
\name{pc_qc_scatter}
\alias{pc_qc_scatter}
\title{Phenocycler QC Scatterplot}
\usage{
pc_qc_scatter(
  df,
  ch_col = "Channel",
  ch_id = NULL,
  int_col,
  id_col = "Code",
  id_samp
)
}
\arguments{
\item{df}{Input data frame containing signal intensities.}

\item{ch_col}{Channel column; Used to
specify the x-axis and can alternatively plot a different variable if
individual channel intensities are not available.}

\item{ch_id}{If plotting an individual channel, which channel should
be plotted?}

\item{int_col}{Intensity column to use.}

\item{id_col}{ID column containing sample information.}

\item{id_samp}{Specific sample ID to plot.}
}
\value{
Boxplot of individual channel or summary of signal intensities.
}
\description{
Generates scatterplot of signal intensity; Must have
X and Y coordinates available in the input dataframe.
}
\examples{

# pheno_qc_scatter <- pc_qc_scatter(
#   df = qc2_filt,
#   ch_id = "DAPI",
#   int_col = "value.z",
#   id_samp = "H10687"
# )

}
