% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.6.pc.img.R
\name{pc_scatter}
\alias{pc_scatter}
\title{Phenocycler Cell Map}
\usage{
pc_scatter(so, type = "var", md_var, col1 = col_grad(scm = 4))
}
\arguments{
\item{so}{Input Seurat object containing coordinates and meta data.}

\item{type}{Scatterplot type ("int" or "var").}

\item{md_var}{If type is "int," specify channel name.}

\item{col1}{Gradient color column to use.}
}
\value{
Cell map scatter plot for selected variable.
}
\description{
Generates phenocycler image scatterplot from Seurat object; Must have
X and Y coordinates available in the input dataframe.
}
\examples{

# psc <- pc_scatter(
#   so = d1,
#   md_var = "cluster"
# )

}
