% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.lib.R
\name{pc_theme_img}
\alias{pc_theme_img}
\title{PC Image Plot Theme}
\usage{
pc_theme_img()
}
\value{
ggplot2 theme parameters to replace default image plotting theme.
}
\description{
Default Phenocycler image plotting theme.
}
\examples{

# pc_theme_img()

}
