% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.check.lib.R
\name{sc.check.lib}
\alias{sc.check.lib}
\title{Install Libraries}
\usage{
sc.check.lib()
}
\value{
Summary of installed packages
}
\description{
Installs CRAN and Bioconductor packages (if not already installed) necessary to run package
}
\examples{

 # sc.check.lib()

}
