% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.DGEA.R
\name{sc.DGEA}
\alias{sc.DGEA}
\title{scRNA-Seq DGEA}
\usage{
sc.DGEA(so, md.list, ct, MAST.comp, MAST.name, form1, parl, core.perc)
}
\arguments{
\item{so}{An object of class Seurat. Must contain the columns 'CellType' and 'CellGroup' in the metadata slot.}

\item{md.list}{A vector of character strings indicating metadata columns for overlaying on a loadings plot.}

\item{ct}{Character string vector containing the name(s) of up to 2 cell type groups present in the CellGroup column.
The provided Seurat object is then subsetted to include only clusters from the provided group names.}

\item{MAST.comp}{Character string indicating the name of the MAST group comparison for conducting DGEA. MAST names are comprised of the chosen
variable name and the leading factor level within that variable.}

\item{MAST.name}{User-defined name of a DGEA comparison, given as a character string.}

\item{form1}{Formula to use for MAST generalized linear model.}

\item{parl}{Logical indicating whether processing should be run in parallel (Linux and WSL2 only). Set to FALSE if running sequentially.}

\item{core.perc}{Percentage of available cores to use if running in parallel (Linux and WSL2 only). Set to 1 if running sequentially.}
}
\value{
A list of DGEA results per cell type for the chosen group comparison, including genes missing fold changes and cell type DGEA results
with errors.
}
\description{
Performs DGEA per cell type of a Seurat Object.
}
\examples{

# dgea.output <- sc.DGEA(
# # Seurat object
# d1,
# # metadata column list
# c(list.p.cols,"CellType","nFeature_RNA"),
# # Cell type column name
# "CellType",
# # MAST comparison name
# "KnockoutKO",
# # MAST name (user-provided)
# "KO vs. NG ",
# # Formula
# as.formula(
#   paste(
#     "~","Knockout","+",
#     "nFeature_RNA",
#     sep = " "
#     )
#   ),
# # run in parallel? (Set to FALSE if on Windows)
# TRUE,
# # core percentage to use
# 0.5
# )

}
