% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.4.integrate.files.R
\name{sc.integrate.data}
\alias{sc.integrate.data}
\title{Batch Integration of scRNA-Seq Datasets}
\usage{
sc.integrate.data(l.so, parl, core.perc)
}
\arguments{
\item{l.so}{List of processed Seurat objects to be integrated.}

\item{parl}{Logical indicating whether processing should be run in parallel (Linux and WSL2 only). Set to FALSE if running sequentially.}

\item{core.perc}{Proportion (as a numeric value) of available cores to use if running in parallel (Linux and WSL2 only). Set to 1 if running sequentially.}
}
\value{
A Seurat object containing integrated data for all samples present in a scRNA-Seq experiment.
}
\description{
Integrates a list of samples processed as Seurat objects.
}
\examples{

# d.integrated <- sc.integrate.data(
#    list.data,
#    TRUE,
#    0.5
#    )

}
