% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.pca.R
\name{sc.pca.plot}
\alias{sc.pca.plot}
\title{scRNA-Seq PCA Loadings Plot}
\usage{
sc.pca.plot(so, md.list)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{md.list}{A vector of character strings indicating metadata columns for overlaying on a loadings plot.}
}
\value{
A series of loadings plots with specified metadata overlays.
}
\description{
Generates a panel of loadings plots given a Seurat object containing PCA results.
}
\examples{

p.pca <- sc.pca.plot(d.integrated,c("col1","col2","col3"))

}
