% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.enrichment.R
\name{sc.plot.enrichment}
\alias{sc.plot.enrichment}
\title{Enrichment Plots}
\usage{
sc.plot.enrichment(l.enr, en.type)
}
\arguments{
\item{l.enr}{A data frame containing enrichment results generated by sc.enrichment().}

\item{en.type}{Are the provided enrichment results based on the Gene Ontology database or a custom gene set? Type "GO"
for Gene Ontology Results and "cstm" for custom gene set enrichment results.}
}
\value{
An enrichment plot depicting the most significant gene sets for a given comparison and cell type.
}
\description{
Plots the top 10 terms from a provided set of enrichment results.
}
\examples{

# # Custom gene set example
# sc.plot.enrichment(
#   enr[["Results"]][enr[["Results"]][["CellType"]] == "2.Secretory",],
#   "cstm"
# )
# # Gene ontology example
# sc.plot.enrichment(
#   enr[enr[["Description"]] == "GO_BP_2.Secretory",],
#   "GO"
# )
# 

}
