% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R
\name{sc.plot.rho}
\alias{sc.plot.rho}
\title{Plot Sample Contamination Fraction}
\usage{
sc.plot.rho(df.par)
}
\arguments{
\item{df.par}{Data frame of updated parameters used for data processing.}
}
\value{
A scatter plot indicating the individual and average contamination fractions for all samples.
}
\description{
Generates a scatter plot indicating the individual and average contamination fractions for all samples.
}
\examples{

sc.plot.rho(list.params)

}
