% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.3.celltype.predictions.R
\name{sc.predict.clusters}
\alias{sc.predict.clusters}
\title{Cell Type Prediction}
\usage{
sc.predict.clusters(so, d.ref, f.size, cl.var, md.list, parl, core.perc)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{d.ref}{An annotated Seurat object containing cell type identities for cell type prediction.}

\item{f.size}{Memory allocated for objects created during cell type prediction (in MB), provided as a numeric value. For larger datasets, set to 5000 or higher.}

\item{cl.var}{Character string containing the name of the cluster variable for cell type predictions.}

\item{md.list}{A vector of character strings indicating metadata columns to include with the prediction score output table.}

\item{parl}{Logical indicating whether processing should be run in parallel (Linux and WSL2 only). Set to FALSE if running sequentially.}

\item{core.perc}{Percentage of available cores to use if running in parallel (Linux and WSL2 only). Set to 1 if running sequentially.}
}
\value{
A list containing a table of marker genes per cluster and an annotated Seurat object with putative cell identities.
}
\description{
Predicts cell type identities of clusters in a Seurat object given a previously annotated data set. The reference data
must have a cell type column named 'CellType' for predictions to be successful.
}
\examples{

# sc.predict.clusters(d.seurat,d.ref,5000,"seurat_clusters",c("col1","col2","col3"),TRUE,0.5)

}
