% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R
\name{sc.process.batch}
\alias{sc.process.batch}
\title{Batch Processing of CellRanger Files}
\usage{
sc.process.batch(df.par, rho.adj, m.cell, m.feat, parl, core.perc)
}
\arguments{
\item{df.par}{Data frame of parameters to use for data processing.}

\item{rho.adj}{Numeric value indicating a proportion to scale rho values calculated during the ambient RNA removal step. 10% (0.1)
is generally suitable for most data sets.}

\item{m.cell}{Threshold for including features in a Seurat Object. Run ?Seurat::CreateSeuratObject() for more details.}

\item{m.feat}{Threshold for including cells in a Seurat Object. Run ?Seurat::CreateSeuratObject() for more details.}

\item{parl}{Logical indicating whether processing should be run in parallel (Linux and WSL2 only). Set to FALSE if running sequentially.}

\item{core.perc}{Percentage of available cores to use if running in parallel (Linux and WSL2 only). Set to 1 if running sequentially.}
}
\value{
A processed list of sample files converted into Seurat objects with a summary list of QC and processing details.
}
\description{
Processes a list of samples into separate Seurat objects for data integration.
}
\examples{

# list.data <- sc.process.batch(
# # parameter list
# list.params,
# # adj.rho proportion
# 0.1,
# # minimum cells per feature
# 5,
# # minimum features per cell
# 200
# )

}
