% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R, R/1.4.integrate.files.R
\name{sc.process.file}
\alias{sc.process.file}
\title{Process CellRanger Data Files}
\usage{
sc.process.file(df.par, i, rho.adj, m.cell, m.feat)

sc.process.file(df.par, i, rho.adj, m.cell, m.feat)
}
\arguments{
\item{df.par}{Character string indicating the path to a set of CellRanger files for data processing.}

\item{i}{A data frame containing sample metadata variables specific to an individual study.}

\item{rho.adj}{Numeric value indicating a proportion to scale rho values calculated during the ambient RNA removal step. 10% (0.1)
is generally suitable for most data sets.}

\item{m.cell}{Threshold for including features in a Seurat Object. See ?Seurat::CreateSeuratObject() for more details.}

\item{m.feat}{Threshold for including cells in a Seurat Object. See ?Seurat::CreateSeuratObject() for more details.}
}
\value{
A processed sample file converted into a Seurat object with a summary list of QC and processing details.

Data frame containing a list of parameters to use for scRNA-Seq data processing by Seurat.
}
\description{
Processes a single sample into a Seurat object for data integration.

Processes a single sample into a Seurat object for data integration.
}
\examples{

proc.data <- sc.process.file(
# parameter list
list.params,
# sample ID
1,
# adj.rho proportion
0.1,
# minimum cells per feature
5,
# minimum features per cell
200
)


# Dataset input parameters
list.params <- Create.proc.param(
  "data/",
  data.frame(
    # ID column name (splits by underscore, should be listed first in folder name as in 's01_1_KO_a')
    "Code" = unlist(lapply(strsplit(basename(list.files("data/")),"_",fixed = T),"[",1)),
    # 1st metadata column (include in CellRanger folder name)
    "Knockout" = as.factor(ifelse(grepl("NG",basename(list.files("data/"))),"ctrl","KO")),
    # 2nd metadata column
    "Region" = as.factor(ifelse(grepl("LAE",basename(list.files("data/"))),"1","2")),
    # 3rd metadata column (add/remove columns as needed)
    "Time" = as.factor(ifelse(grepl("D28",basename(list.files("data/"))),"a","b"))
  )
)

}
