% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc.top10.deg.dotplot}
\alias{sc.top10.deg.dotplot}
\title{Top-10 DEG Dot Plot}
\usage{
sc.top10.deg.dotplot(p.type, l.deg, so, ct, cl.var, list.var)
}
\arguments{
\item{p.type}{Should a custom gene list or DGEA results object be used for plotting? Type "cstm.list" for custom gene lists and "deg.list"
for using dgea.results objects.}

\item{l.deg}{A list of DGEA results returned by sc.DGEA() or a vector of selected genes for plotting.}

\item{so}{An object of class Seurat.}

\item{ct}{A pattern provided as a character string for matching a specific cell type or types.}

\item{cl.var}{Character string indicating the name of the cluster/cell type variable.}

\item{list.var}{A vector of character strings indicating the name(s) of up to two group variables for stratifying plot points.}
}
\value{
An input data frame and corresponding dot plot displaying the expression of the top-10 DEGs for a specific cell type.
}
\description{
Generates a dot plot from a Seurat Object and DEG list based on the top-10 DEGs for a specific cell type and comparison.
}
\examples{

p.dotplot <- sc.top10.deg.dotplot(
# Type "deg.list" or "cstm.list" to toggle between inputs
"deg.list",
# Name of a custom gene list or dgea.results object
dgea.output,
# Seurat object
d.annotated,
# Unique character strings corresponding to cell types
"3.Se|6.Se",
# Name of clustering variable
"CellType",
# Vector of up to 2 variables for stratifying clustering variables
c("Knockout","Airway")
)

}
