% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc.top10.marker.heatmap}
\alias{sc.top10.marker.heatmap}
\title{Top-10 Marker Gene Heatmap}
\usage{
sc.top10.marker.heatmap(so, cl.var, h.w, h.h, fs.c, fs.r)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{cl.var}{Character string containing the name of the cluster variable for cell type predictions.}

\item{h.w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h.h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs.c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs.r}{Numeric value for row fontsize (passed to ComplexHeatmap).}
}
\value{
A ComplexHeatmap object containing a top-10 marker gene heatmap.
}
\description{
Generates a heatmap from a Seurat Object and marker gene list based on the top-10 marker genes for each cluster.
}
\examples{

# p.umap <- sc.top10.marker.heatmap(d.annotated,"seurat.clusters",18,24,6,8)

}
