% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc.umap.panel.gene.list}
\alias{sc.umap.panel.gene.list}
\title{Visualize Gene List Expression}
\usage{
sc.umap.panel.gene.list(
  list.g,
  so,
  md.var,
  g.name,
  col.scheme,
  col.names,
  leg.x,
  leg.y
)
}
\arguments{
\item{list.g}{A vector containing a list of genes to be plotted for a given Seurat object.}

\item{so}{An object of class Seurat.}

\item{md.var}{A character string indicating the clustering column for overlaying on a UMAP plot.}

\item{col.scheme}{The color scheme to be used for distinguishing between groups, provided as a vector.}

\item{col.names}{A vector of the same length as the provided color scheme for assigning colors to each group.}

\item{leg.x}{A numeric value indicating the placement of the figure legend on the x-axis.}

\item{leg.y}{A numeric value indicating the placement of the figure legend on the y-axis.}
}
\value{
A list of plots saved as ggplot2 objects for visualizing cluster gene expression.
}
\description{
Generates a series of plots to visualize the expression of a list of genes per cluster.
}
\examples{

sc.umap.panel.gene.list(list.genes,d.seurat,"seurat_clusters",col.vec,col.vec.names,0.95,0.95)

}
