% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc.umap.panel}
\alias{sc.umap.panel}
\title{scRNA-Seq UMAP Plot}
\usage{
sc.umap.panel(so, md.list)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{md.list}{A vector of character strings indicating metadata columns for overlaying on a loadings plot.}
}
\value{
A series of UMAPs with specified metadata overlays.
}
\description{
Generates a panel of UMAPs given a Seurat object containing PCA and UMAP results.
}
\examples{

# p.umap <- sc.umap.panel(d.integrated,c("col1","col2","col3"))

}
