% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.volplot.R
\name{sc.volcano}
\alias{sc.volcano}
\title{Volcano Plot}
\usage{
sc.volcano(l.deg, ct, comp.name, p.cut, f.cut, f.lim, y.limit)
}
\arguments{
\item{l.deg}{A list of DGEA results returned by sc.DGEA().}

\item{ct}{A pattern provided as a character string for matching a specific cell type.}

\item{comp.name}{comparison name specified by the user and returned by sc.DGEA, provided as a character string.}

\item{p.cut}{Numeric value for p-value cutoff to indicate significance.}

\item{f.cut}{Numeric value for fold change cutoff to indicate a high effect size.}

\item{f.lim}{Numeric value for fold change limits on the x-axis (given in multiples of 6).}

\item{y.limit}{Numeric value for the upper-bound -log10 p-value to display on the y-axis.}
}
\value{
A volcano plot for the chosen cell type and treatment comparison.
}
\description{
Generates a volcano plot from a DEG results object for a chosen cell type and comparison.
}
\examples{

# p.vol <- sc.volcano(dgea.output,"1.Secretory","KO vs. Control",0.005,0.25,6)

}
