% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.4.coverage.plot.r
\name{sc_atac_motifs}
\alias{sc_atac_motifs}
\title{Add JASPAR TF Motifs}
\usage{
sc_atac_motifs(so, dref)
}
\arguments{
\item{so}{An object of class Seurat. Must contain an ATAC assay.}

\item{dref}{Path to a .gtf file containing reference gene annotations.}
}
\value{
An annotated ChromatinAssay object.
}
\description{
Creates a ChromatinAssay from a scATAC-Seq counts assay
and annotates transcription factor motifs based on
the JASPAR database.
}
\examples{

# d_chrmasy <- sc_atac_motifs(
#   d,
#   rtracklayer::import(
#    "ref/gencode.v45.primary_assembly.annotation.gtf"
#    )
# )

}
