% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.5.cellchat.R
\name{sc_cc_chrd}
\alias{sc_cc_chrd}
\title{CellChat Chord Diagram}
\usage{
sc_cc_chrd(
  ccdf,
  title1,
  title2,
  cc_type,
  g_name = NULL,
  pw_sel = NULL,
  pw,
  ph,
  fs1,
  fd1,
  lgx = 0.1,
  lgy = 0.1,
  g_ind,
  list_ct
)
}
\arguments{
\item{ccdf}{A CellChat data frame generat.}

\item{title1}{Plot title, provided as a character string.}

\item{cc_type}{Type of plot to use (either "total" of "comp"). Selecting
"total" plots the total number of interactions between cell types, whereas
"comp" compares the fold difference between two treatment groups}

\item{g_name}{Group name to split data by if cc_type is "comp."}

\item{pw_sel}{Character string of a pathway to use for filtering the
provided CellChat object.}

\item{pw}{Plot width (in cm).}

\item{ph}{Plot height (in cm).}

\item{fs1}{Font size (between 0.1 and 1).}

\item{fd1}{Gap between cell type labels and plot (generally between 1 and 3).}
}
\value{
A chord diagram displaying either pathway or individual
ligand-receptor interactions.
}
\description{
Creates a chord diagram from a provided CellChat results data frame.
}
\examples{

# pchrd <- sc_cc_chrd(
#   ccdf = ccm[[1]],
#   title1 = "CellChat: Total L-R Interactions",
#   cc_type = "total",
#   pw = 12,
#   ph = 12,
#   fs1 = 0.6,
#   fd1 = 2.8
# )

}
