% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.check.lib.R
\name{sc_check_lib}
\alias{sc_check_lib}
\title{Install Libraries}
\usage{
sc_check_lib()
}
\value{
Summary of installed packages
}
\description{
Installs CRAN and Bioconductor packages (if not already installed)
}
\examples{

 # sc_check_lib()

}
