% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc_top10_deg_dotplot}
\alias{sc_top10_deg_dotplot}
\title{Top-10 DEG Dot Plot}
\usage{
sc_top10_deg_dotplot(
  p_type,
  l_deg,
  so,
  ct,
  cl_var,
  split_var,
  list_var,
  col1,
  vline1
)
}
\arguments{
\item{p_type}{Should a custom gene list or DGEA results object
be used for plotting?
Type "cstm.list" for custom gene lists and "deg.list"
for using dgea.results objects.}

\item{l_deg}{A list of DGEA results returned by sc.DGEA()
or a vector of selected genes for plotting.}

\item{so}{An object of class Seurat.}

\item{ct}{A pattern provided as a character string for
matching a specific cell type or types.}

\item{cl_var}{Character string indicating the name of the
cluster/cell type variable.}

\item{split_var}{Logical indicating whether the cluster variable
should be stratified by additional group variables.}

\item{list_var}{(Optional) A vector of character strings
indicating the name(s) of up to two group variables
for stratifying plot points.}

\item{col1}{Gradient color scheme to use.}

\item{vline1}{Add a vertical line to plot?}
}
\value{
An input data frame and corresponding dot plot displaying
the expression of the top-10 DEGs for a specific cell type.
}
\description{
Generates a dot plot from a Seurat Object and DEG list based on
the top-10 DEGs for a specific cell type and comparison.
}
\examples{

# p_dotplot <- sc_top10_deg_dotplot(
#   # Type "deg.list" or "cstm.list" to toggle between inputs
#   "deg.list",
#   # Name of a custom gene list or dgea.results object
#   dgea_output,
#   # Seurat object
#   d_annotated,
#   # Unique character strings corresponding to cell types
#   "3.Se|6.Se",
#   # Name of clustering variable
#   "CellType",
#   TRUE,
#   # Vector of up to 2 variables for stratifying clustering variables
#   c("Knockout","Airway")
# )

}
