% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc_top10_marker_heatmap_rc}
\alias{sc_top10_marker_heatmap_rc}
\title{Top-10 Marker Gene Heatmap (Reclustered)}
\usage{
sc_top10_marker_heatmap_rc(so, cl_var, h_w, h_h, fs_c, fs_r)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{cl_var}{Character string containing the name of
the cluster variable for cell type predictions.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}
}
\value{
A ComplexHeatmap object containing a reclustered
top-10 marker gene heatmap.
}
\description{
Generates a heatmap from a reclustered Seurat Object
and marker gene list based on the top-10 marker genes
for each cluster.
}
\examples{

# p_umap <- sc_top10_marker_heatmap(d_annotated,"seurat_clusters",18,24,6,8)

}
