% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1.heatmaps.R
\name{sc_top_motif_heatmap}
\alias{sc_top_motif_heatmap}
\title{Top Differentially Active Motif Heatmap}
\usage{
sc_top_motif_heatmap(
  so,
  mot_m,
  filt_ct,
  ct_name,
  cl_var,
  top_n,
  h_w,
  h_h,
  fs_c,
  fs_r
)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{mot_m}{Character string providing the name of the assay to use.}

\item{filt_ct}{Logical indicating whether the data should
be filtered based on cell type}

\item{ct_name}{Character string containing the cell type name to filter.}

\item{cl_var}{Character string containing the name
of the clustering variable.}

\item{top_n}{Character string containing the name of the clustering variable.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}
}
\value{
A ComplexHeatmap object containing a top motif heatmap.
}
\description{
Generates a heatmap from a Seurat Object and
differential activity analysis result.
}
\examples{

# tf_heatmap <- sc_top_motif_heatmap(
#   # Seurat object
#   so = d,
#   # Differential activity results
#   mot_m = diff_output_activity,
#   # Filter based on CellType?
#   filt_ct = TRUE,
#   # Cell type name
#   ct_name = "11.Secretory",
#   # Clustering column
#   cl_var = "CellType",
#   # Number of motifs to use
#   top_n = 50,
#   # Heatmap width
#   h_w = 36,
#   # Heatmap height
#   h_h = 12,
#   # Column font size
#   fs_c = 6,
#   # Row font size
#   fs_r = 8
# )

}
