% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2.umap.R
\name{sc_vio_panel_gene}
\alias{sc_vio_panel_gene}
\title{Visualize Individual Gene Expression}
\usage{
sc_vio_panel_gene(
  so,
  md_var,
  g_name,
  col_scheme,
  col_names,
  leg_x,
  leg_y,
  slot1
)
}
\arguments{
\item{so}{An object of class Seurat.}

\item{md_var}{A character string indicating
the clustering column for overlaying on a UMAP plot.}

\item{g_name}{A character string indicating the gene name.}

\item{col_scheme}{The color scheme to be used for
distinguishing between groups, provided as a vector.}

\item{col_names}{A vector of the same length as the
provided color scheme for assigning colors to each group.}

\item{leg_x}{A numeric value indicating the placement
of the figure legend on the x-axis.}

\item{leg_y}{A numeric value indicating the placement
of the figure legend on the y-axis.}

\item{slot1}{A character string corresponding to the umap slot name to plot.}
}
\value{
A series of plots stored as a ggplot2 object
for visualizing cluster gene expression.
}
\description{
Generates a series of plots to visualize
individual gene expression per cluster.
}
\examples{

# p_umap <- sc_umap_panel_gene(
#  d_integrated,
#  c("col1","col2","col3"),
#  "CFTR",
#  col_univ,
#  c("group1","group2"),
#  0.95,
#  0.95,
#  "wnn.umap"
# )

}
