% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1.pca.R
\name{sc.pca}
\alias{sc.pca}
\title{Seurat PCA}
\usage{
sc.pca(so)
}
\arguments{
\item{so}{An object on class Seurat.}
}
\value{
A Seurat object containing PCA results.
}
\description{
Performs PCA on a Seurat object.
}
\examples{

d.integrated <- sc.pca(d.integrated)

}
