% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3.process.files.R
\name{sc.process.file}
\alias{sc.process.file}
\title{Process CellRanger Data Files}
\usage{
sc.process.file(df.par, i, rho.adj, m.cell, m.feat)
}
\arguments{
\item{df.par}{Data frame of parameters to use for data processing.}

\item{i}{Numeric value for sample ID.}

\item{rho.adj}{Numeric value indicating a proportion to scale rho values calculated during the ambient RNA removal step. 10% (0.1)
is generally suitable for most data sets.}

\item{m.cell}{Threshold for including features in a Seurat Object. Run ?Seurat::CreateSeuratObject() for more details.}

\item{m.feat}{Threshold for including cells in a Seurat Object. Run ?Seurat::CreateSeuratObject() for more details.}
}
\value{
A processed sample file converted into a Seurat object with a summary list of QC and processing details.
}
\description{
Processes a single sample into a Seurat object for data integration.
}
\examples{

# proc.data <- sc.process.file(
# # parameter list
# list.params,
# # sample ID
# 1,
# # adj.rho proportion
# 0.1,
# # minimum cells per feature
# 5,
# # minimum features per cell
# 200
# )

}
